requirements.txt
gradescope-utils>=0.5.0
pandas>=2.2.2
gitpython>=3.1.43